refactor InitUseCase to SplashActivityStrategy
