+++ app/src/main/java/org/eyeseetea/malariacare/strategies/ASplashActivityStrategy.java
@@ -1,23 +1,20 @@
-package org.eyeseetea.malariacare.domain.usecase;
+package org.eyeseetea.malariacare.strategies;
 
 import android.app.Activity;
 import android.content.Intent;
 
-/**
- * Created by idelcano on 12/12/2016.
- */
 
-public abstract class AInitUseCase {
+public abstract class ASplashActivityStrategy {
 
     protected Activity activity;
 
-    public AInitUseCase(Activity activity) {
+    public ASplashActivityStrategy(Activity activity) {
         this.activity = activity;
     }
 
     public abstract void finishAndGo();
 
-    public void finishAndGo(Class<? extends Activity> activityClass) {
+    protected void finishAndGo(Class<? extends Activity> activityClass) {
         activity.startActivity(new Intent(activity, activityClass));
         activity.finish();
     }
